refactor(architecture): infrastructure -> model 허용으로 전환하고 JPA Adapter 를 infrastructure 로 복귀
- CLAUDE.md 의존 규칙을 Inbound(presentation -> domain) / Outbound(domain -> infrastructure -> model) 로 재정의
- 기존 `infrastructure -> model = X` 규칙은 JPA Adapter 가 Entity/Repository 를 직접 다뤄야 하는 현실과 충돌했음
- JpaRefreshTokenStore(+Test) 를 domain/auth/service 에서 infrastructure/auth/jpa 로 이동해 Port/Adapter 분리를 정상화
